Rework how term views work.
Write a typechecking pass for STLC.  My manual handling of De Bruijn
indices is starting to be a real problem.  The typechecker, in particular
is really fiddly.  Minor changes that I think should be OK in fact
break things.

I'm going to investigate other ways to handle the binding issue.