Enhance weakenings.
Weakenings can now be applied under variables, and the Weak constructor
may have an entire weakening inside it.  This simplifies the user API
because weakenings do not need to be exposed when using views on
binders (VLam, VPi, etc).  It complicates somewhat the interaction of
substituion with weakenings; however, the resulting system is better,
I think.