Partial impelementation of row polymorphism.
This commit adds a new type of row value which can be abstracted
over and manipulated.  The shape of a record type is determined
by a row.  In this commit, row values are first-class, but labels
are not.

I'm reconsidering this whole line of work, for two major reasons.
First, I think I accidentily broke canonical forms.  Rows basically
give you a back-door way to abstract over types (the whole point, really)
and one has to be very careful to ensure that terms remain eta-long
when you do this.  The problem occurs when you have a term variable
which is typed by a type variable.  Maybe it all works out, but I'll
have to think carefully about it.

Second, I don't think this system actually solves the problem I care
about in the end.  I've just about convinced myself that the encoding
I have in mind for mutually-recursive let actually requires first-class
labels, and that seems like a bridge too far.  The metatheory of the calculus
becomes a lot more complicated with first-class labels.

If this line of work continues, I need to:
  1) Update term views to accomidate the new forms
  2) Convince myself I haven't destroyed the metatheory
  3) Extend unification to handle records and rows