Add hashing support and use it to short-cut alpha-eq tests in DAG implementation.
Still not entirely happy with the alpha-equivalance test, as it doesn't short-cut
in cases where two identical terms are applied to distinct weakenings.  This is
because it might happen that two different weakenings nonetheless send equal
terms to equal terms.  It seems that we should be able to find an accurate
test based on the free variable sets of the two terms, but I haven't worked out
what it should be, yet.